+++ src/verify.h
@@ -10,14 +10,14 @@ extern "C" {
 
 int verify(libsign_public_key *public_key, libsign_signature *signature, const char *filename);
 int verify_buffer(libsign_public_key *public_key, libsign_signature *signature,
-                  const uint8_t *data, uint64_t datalen);
+                  const uint8_t *data, uint32_t datalen);
 
 int rsa_sha1_verify_file(libsign_public_key *pub_ctx, libsign_signature *sig_ctx,
                          const char *filename);
 int rsa_sha1_verify_fd(libsign_public_key *pub_ctx, libsign_signature *sig_ctx,
                           int fd);
 int rsa_sha1_verify_data(libsign_public_key *pub_ctx, libsign_signature *sig_ctx,
-                          const uint8_t *data, uint64_t datalen);
+                          const uint8_t *data, uint32_t datalen);
 
 #ifdef __cplusplus
 }
